571: refactor methods and types
