--- app/db/db_types.py
@@ -1,10 +1,0 @@
-
-
-
-
-
-
-# You shouldn't create or change this explicitly
-
-
-
