+++ app/db/methods/criteria_pack.py
@@ -1,11 +1,9 @@
 from datetime import datetime
 
-from app.db.db_methods import get_criteria_pack_collection
-from app.db.methods.client import get_client, get_db, get_fs
+from app.db.db_main import get_criteria_pack_collection
+from app.db.methods.client import get_db, get_fs
 
-# from app.db.types. import Logs
 
-# client = get_client()
 db = get_db()
 fs = get_fs()
